Refactor functions into different file for clarity
